api/requirements.txt
-i https://pypi.org/simple
flask==1.1.2
matplotlib==3.3.3
numpy==1.18.5
pillow==8.0.1; python_version >= '3.6'
pymongo==3.11.0
pyyaml==5.3.1